Define SETTINGS_BUFLEN [ESD-1028] (#9)
* Define SETTINGS_BUFLEN and use it

* Use size_t for buffer len argument type in public API

* Update version number to 0.1.6

+++ include/libsettings/settings_util.h
@@ -36,7 +36,7 @@ int settings_format(const char *section,
                     const char *value,
                     const char *type,
                     char *buf,
-                    uint16_t blen);
+                    size_t blen);
 
 /**
  * @brief   Parse setting strings from SBP message buffer
@@ -57,7 +57,7 @@ int settings_format(const char *section,
  * @retval -1               An error occurred.
  */
 settings_tokens_t settings_parse(const char *buf,
-                                 uint16_t blen,
+                                 size_t blen,
                                  const char **section,
                                  const char **name,
                                  const char **value,
